requirements.txt
# Core
Django==5.0
django-configurations==2.5
pytz==2023.3.post1
python-dotenv==1.0.0
gunicorn==21.2.0

# Database
psycopg2-binary==2.9.9
django-allauth==0.58.2

# Rest apis
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
social-auth-app-django
django-filter==23.5
django-cors-headers==4.3.1

# Test
pytest-django==4.8.0
pytest==8.0.0

# Celery
celery[redis]==5.3.6